% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHLloess.R
\name{NHLloess}
\alias{NHLloess}
\title{Locally weighted regression smoothing}
\usage{
NHLloess(model, alpha = 0.05)
}
\arguments{
\item{model}{a dataframe, specifically, ouput from the function \code{NHLmodel}.}

\item{alpha}{the smoothing parameter; see \code{?loess}}
}
\value{
A NHLmodel dataframe with an additional column - \code{loess} the local regression values.
}
\description{
Wrapper function for computing locally weigthed regression smoothing of NHL shot data, for comparison with the output NHL model values.
}
\details{
Will be expanded later on.
}
