% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHLmodel.R
\name{NHLmodel}
\alias{NHLmodel}
\title{Spatial model for NHL goal scoring probability}
\usage{
NHLmodel(x, ngibbs, burnin, p.max, inits, outfile)
}
\arguments{
\item{x}{a dataframe, specifically, output from the function \code{shots2model}.}

\item{ngibbs}{numeric, number of Gibbs samples to run the model for.}

\item{burnin}{numeric, number of Gibbs samples to discard as burn-in.}

\item{p.max}{numeric, tuning parameter indicating the maximum probability value at the goal (should be between 0 and 1).}

\item{inits}{a dataframe containing values to initiate the Gibbs sampler (see details).}

\item{outfile}{output path and filename (.RData) for storing the model results (8 objects; see details).}
}
\value{
The function returns a dataframe with the posterior mean and variance of the model which can be used to map spatial variation in NHL goal scoring probabilities.
}
\description{
Calculate spatial probability of NHL goal scoring based on shot location data.
}
\details{
Will be expanded later on.
}
