% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHLmodelplot.R
\name{NHLmodelplot}
\alias{NHLmodelplot}
\title{Plot NHL model results}
\usage{
NHLmodelplot(model, plot = "none")
}
\arguments{
\item{model}{a dataframe, specifically, ouput from the function \code{NHLmodel}.}

\item{plot}{character, what to plot; one of (\code{'shots','goals','postmean','postvar','none'}). If none is given just the rink outline is plotted.}
}
\value{
A half-rink plot of shots or goals (as points) or posterior mean or variance (as a surface).
}
\description{
Very basic plotting function for outputs from the spatial model of NHL goal scoring probabilities.
}
\details{
Will be expanded later on.
}
