% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlmodelr-package.R
\docType{package}
\name{nhlmodelr-package}
\alias{nhlmodelr-package}
\title{wildlifeTG - Time Geographic Analysis of Wildlife Telemetry Data}
\description{
The package \code{nhlmodelr} is for spatial modelling of NHL shot location data.
}
\details{
The package uses a model that has been developed by Jed Long and Time Swartz.
}
\author{
Jed Long
}
