% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shots2model.R
\name{shots2model}
\alias{shots2model}
\title{Format shot data for model}
\usage{
shots2model(shots)
}
\arguments{
\item{shots}{a dataframe, specifically, a shots data frame downloaded from the package \code{nhlmodelr}.}
}
\value{
The function returns a dataframe with four columns: x and y locations, and number of shots and goals from each location.
}
\description{
Format shot location data for spatial model for NHL goal scoring probability
}
\details{
Will be expanded later on.
 - Automatically removes penalty shots and empty net situations.
 - Currently only considers reg. season games, could be updated.
}

