% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamInfo.R
\docType{data}
\name{teamInfo}
\alias{teamInfo}
\title{Team information data}
\format{An \code{data.frame} object providing basic linking info for each team ID.}
\description{
Team information data to be linked to team ID.
}
\details{
Basic team information data, really to be used as an internal look up table for certain functions.
}
\keyword{datasets}
