% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{teams}
\alias{teams}
\title{Team shooting analysis}
\usage{
teams(model, shots)
}
\arguments{
\item{model}{a dataframe, specifically, ouput from the function \code{NHLmodel}.}

\item{shots}{the raw data shots file that was used as input into the function \code{shots2model}.}
}
\value{
The function returns a dataframe with spatially adjusted team shooting statistics.
}
\description{
Calculate spatially adjusted team-based shooting  statistics using spatially explicit model of NHL scoring probability.
}
\details{
Will be expanded later on.
}
